requirements.txt
fastapi
flask
uvicorn
python-dotenv
faiss-cpu
numpy
torch
transformers
huggingface-hub
